refactor(api): minor refactor and linting
