+++ apps/atqr/api/src/app/app.module.ts
@@ -4,7 +4,8 @@ import { ConfigModule, ConfigService } from '@nestjs/config';
 import { ChallengeController } from './challenge.controller';
 
 import { PrismaService } from './infra/database/prisma.service';
-import { Mailer } from './infra/email/mailer-sevice';
+import { Mailer } from './infra/email/mailer.service';
+
 import { StripeService } from './infra/payment/stripe.service';
 import { ChallengeRepository } from './repositories/challenge.repository';
 import { PlayerRepository } from './repositories/player.repository';
