+++ apps/atqr/api/src/app/challenge.controller.ts
@@ -20,7 +20,7 @@ import { CreateChallengeDto } from './dtos/createChallenge.dto';
 import { UpdateCreditCardTokenDto } from './dtos/updateCreditCardToken.dto';
 import ValidationErrors, {
   ValidationErrorTypes,
-} from './errors/validationErrors';
+} from './errors/validationError';
 import { Mailer } from './infra/email/mailer.service';
 import { ChallengeRepository } from './repositories/challenge.repository';
 import { PlayerRepository } from './repositories/player.repository';
